比較登録処理作成

+++ project/app/OpenAPI/Model/User.php
@@ -58,7 +58,7 @@ class User implements ModelInterface, ArrayAccess, \JsonSerializable
       * @var string[]
       */
     protected static $openAPITypes = [
-        'id' => 'float',
+        'id' => 'int',
         'name' => 'string'
     ];
 
@@ -303,7 +303,7 @@ class User implements ModelInterface, ArrayAccess, \JsonSerializable
     /**
      * Gets id
      *
-     * @return float
+     * @return int
      */
     public function getId()
     {
@@ -313,7 +313,7 @@ class User implements ModelInterface, ArrayAccess, \JsonSerializable
     /**
      * Sets id
      *
-     * @param float $id ユーザーID
+     * @param int $id ユーザーID
      *
      * @return self
      */
